% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsic.R
\name{join}
\alias{join}
\title{Join elements into a string}
\usage{
join(sep, v, ...)
}
\arguments{
\item{sep}{A string used as a separator}

\item{v}{A string vector to be joined}

\item{...}{other variables to be converted to strings and joined}
}
\value{
A joined string
}
\description{
convert elements of vector v (and all following arguments)
in strings and join them using sep as separator.
}
\examples{
join(" ", c("Hello", "World"))
}
