% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsic.R
\name{ldp}
\alias{ldp}
\title{Least Distance Problem}
\usage{
ldp(u, co, rcond = 1e+10)
}
\arguments{
\item{u}{a dense matrix of inequality constraints}

\item{co}{a right hand side vector of inequality constraints}

\item{rcond}{maximal condition number for determining rank deficient matrix}
}
\value{
solution vector or NULL if constraints are unfeasible
}
\description{
Solve least distance programming: find x satisfying \code{u\%*\%x >= co} and s.t. min(||x||)
by passing to nnls (non negative least square) problem.
}
