% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsic.R
\name{ls_0}
\alias{ls_0}
\title{Particular solution of rank-deficient least squares}
\usage{
ls_0(aq, b)
}
\arguments{
\item{aq}{QR decomposition of a matrix}

\item{b}{right-hand-side of the LS system}
}
\value{
solution vector where free variables are set to 0.
}
\description{
Particular solution of rank-deficient least squares
}
