% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsic.R
\name{ls_ln}
\alias{ls_ln}
\title{Linear Least Squares, least norm solution}
\usage{
ls_ln(a, b, rcond = 1e+10)
}
\arguments{
\item{a}{matrix}

\item{b}{vector of right hand side}

\item{rcond}{maximal condition number for rank definition}
}
\description{
Linear Least Squares, least norm solution
}
