% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsic.R
\name{ls_ln_svd}
\alias{ls_ln_svd}
\title{Linear Least Squares, least norm solution (by svd)}
\usage{
ls_ln_svd(a, b, rcond = 1e+10)
}
\arguments{
\item{a}{dense matrix}

\item{b}{right hand side vector}

\item{rcond}{maximal condition number for determining rank deficient matrix}
}
\value{
solution vector
}
\description{
Least squares \code{a\%*\%x ~= b} of least norm ||x|| by using svd(a)
}
