% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsic.R
\name{Nulla}
\alias{Nulla}
\title{Null-space basis}
\usage{
Nulla(M, rcond = 1e+10)
}
\arguments{
\item{M}{matrix such that \code{t(M)\%*\%B=0} where B is a basis of t(M)'s kernel (aka Null-space)}

\item{rcond}{maximal condition number for rank definition}
}
\value{
numeric matrix whose columns are basis vectors. Its attribute 'qr' contains QR decomposition of M.
}
\description{
use Lapack for null space basis
(derived from MASS::Null)
}
\examples{
Nulla(1:3)
}
\seealso{
\link[MASS:Null]{MASS::Null}
}
