% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsic.R
\name{tls}
\alias{tls}
\title{Total Least Squares \code{a\%*\%x ~= b}}
\usage{
tls(a, b)
}
\arguments{
\item{a}{matrix}

\item{b}{right hand side vector}
}
\value{
solution vector
}
\description{
Total Least Squares \code{a\%*\%x ~= b}
}
